# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ISAGI NEWS API",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://isagiapi.galder315.ga/"
    },
    {
      "url": "http://localhost:3000/"
    }
  ],
  "paths": {
    "/account/sales/liked": {
      "get": {
        "tags": ["account", "sales"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de ofertas favoritas del usuario",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SalesResponse"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["account", "sales"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaleIdRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Oferta añadida a favoritos correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Datos no validos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Oferta ya añadida a favoritos o no existe",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["account", "sales"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaleIdRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Oferta eliminada de favoritos correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Datos no validos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Oferta no añadida a favoritos o no existe",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/faq/{id}": {
      "put": {
        "tags": ["admin"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id de la pregunta frecuente",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Faq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Pregunta frecuente editada correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FaqUniqueResponse"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/sales/{id}": {
      "put": {
        "tags": ["admin"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id de la oferta",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SalesUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Oferta actualizada correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SalesUniqueResponse"
                }
              }
            }
          },
          "400": {
            "description": "Datos no validos o oferta no encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado o oferta no encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/users": {
      "get": {
        "tags": ["admin"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de usuarios",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsersAdminResponse"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/users/{id}": {
      "put": {
        "tags": ["admin"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id del usuario",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Usuario actualizado correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserUniqueResponse"
                }
              }
            }
          },
          "400": {
            "description": "Datos no validos o usuario no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado o usuario no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/users/{id}/reset-password": {
      "post": {
        "tags": ["admin"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id del usuario",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contraseña del usuario reseteada correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasswordResetResponse"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Usuario no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/sales/{id}": {
      "get": {
        "tags": ["sales"],
        "security": [
          {
            "BearerAuth": []
          },
          {}
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id de la oferta",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Oferta con el id especificado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SalesUniqueResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "MessageResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "description": "Codigo de estado de la respuesta"
          },
          "message": {
            "type": "string",
            "description": "Mensaje de respuesta"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "description": "Codigo de estado de la respuesta"
          },
          "error": {
            "type": "string",
            "description": "Mensaje de respuesta"
          }
        }
      },
...
      "SalesResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "description": "Codigo de estado de la respuesta"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Sale"
            }
          }
        }
      },
      "SalesUniqueResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "description": "Codigo de estado de la respuesta"
          },
          "data": {
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/Sale"
              }
            ]
          }
        }
      },
      "Sale": {
        "type": "object",
        "properties": {
          "sale_id": {
            "type": "integer",
            "description": "Identificador de la oferta"
          },
          "title": {
            "type": "string",
            "description": "Titulo de la oferta"
          },
          "description": {
            "type": "string",
            "description": "Descripcion de la oferta"
          },
          "image": {
            "type": "string",
            "description": "Enlace a la imagen de la oferta"
          },
          "shop": {
            "type": "string",
            "description": "Tienda de la oferta"
          },
          "old_price": {
            "type": "string",
            "description": "Precio antiguo de la oferta"
          },
          "new_price": {
            "type": "string",
            "description": "Precio nuevo de la oferta"
          },
          "user_id": {
            "type": "integer",
            "description": "Identificador del usuario que ha publicado la oferta"
          },
          "created_at": {
            "type": "string",
            "description": "Fecha de creacion de la oferta"
          },
          "username": {
            "type": "string",
            "description": "Nombre de usuario del usuario que ha publicado la oferta"
          },
          "likes": {
            "type": "integer",
            "description": "Numero de likes de la oferta"
          },
          "comments": {
            "type": "integer",
            "description": "Numero de comentarios de la oferta"
          }
        }
      },
...
      "Faq": {
        "type": "object",
        "properties": {
          "question_id": {
            "type": "integer",
            "description": "Identificador de la pregunta frecuente"
          },
          "question": {
            "type": "string",
            "description": "Pregunta frecuente"
          },
          "answer": {
            "type": "string",
            "description": "Respuesta a la pregunta frecuente"
          }
        }
      },
      "FaqUniqueResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer"
          },
          "data": {
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/Faq"
              }
            ]
          }
        }
      },
...
      "SaleIdRequest": {
        "type": "object",
        "properties": {
          "saleId": {
            "type": "integer",
            "description": "Identificador de la oferta"
          }
        }
      },
...
      "UsersAdminResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "description": "Codigo de estado de la respuesta",
            "example": 200
          },
          "data": {
            "type": "array",
            "items": {
              "properties": {
                "user_id": {
                  "type": "integer",
                  "description": "Identificador del usuario"
                },
                "email": {
                  "type": "string",
                  "description": "Email del usuario"
                },
                "username": {
                  "type": "string",
                  "description": "Nombre de usuario"
                },
                "email_verified_at": {
                  "type": "string",
                  "description": "Fecha de verificacion del email"
                },
                "admin": {
                  "type": "integer",
                  "description": "Si el usuario es administrador"
                },
                "blocked": {
                  "type": "integer",
                  "description": "Si el usuario esta bloqueado"
                }
              }
            }
          }
        }
      },
      "PasswordResetResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "description": "Codigo de estado de la respuesta",
            "example": 200
          },
          "data": {
            "type": "object",
            "properties": {
              "password": {
                "type": "string",
                "description": "Contraseña del usuario"
              }
            }
          }
        }
      },
      "SalesUpdateRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Titulo de la oferta"
          },
          "description": {
            "type": "string",
            "description": "Descripcion de la oferta"
          },
          "shop": {
            "type": "string",
            "description": "Tienda de la oferta"
          },
          "link": {
            "type": "string",
            "description": "Enlace a la oferta"
          },
          "old_price": {
            "type": "number",
            "description": "Precio antiguo de la oferta"
          },
          "new_price": {
            "type": "number",
            "description": "Precio nuevo de la oferta"
          }
        }
      },
      "UserUpdateRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "Nombre de usuario"
          },
          "email": {
            "type": "string",
            "description": "Email del usuario"
          },
          "admin": {
            "type": "integer",
            "description": "Si el usuario es administrador"
          },
          "blocked": {
            "type": "integer",
            "description": "Si el usuario esta bloqueado"
          }
        }
      },
      "UserUniqueResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer"
          },
          "data": {
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/UsersAdminResponse"
              }
            ]
          }
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}